+++ frontend/catalyst/device/decomposition.py
@@ -11,8 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""This module contains the preprocessing functions.
+
 """
+This module contains the decomposition functions to pre-process tapes for
+compilation & execution on devices.
+"""
+
 import jax
 import pennylane as qml
 from pennylane import transform